Feat: 로그인 디버깅, 로그아웃, 회원탈퇴 구현

+++ project/Myo_jib_sa/app/src/main/java/com/example/myo_jib_sa/mypage/api/MypageData.kt
@@ -56,9 +56,9 @@ data class UserMissionResDto(
 )
 
 data class LogoutResponse(
-    val result: Int
+    val result: Long
 ):  BaseResponse()
 
 data class UnregisterResponse(
-    val result: Int
+    val result: Long
 ):  BaseResponse()
